requirements.txt
langgraph
langchain
langchain-openai
openai
streamlit
fpdf2
python-dotenv
